Refactor : 컴포넌트 파일 폴더분류
