+++ src/Router.js
@@ -9,8 +9,8 @@ import LoginTaeyoung from './pages/taeyoung/Login/Login';
 import MainTaeyoung from './pages/taeyoung/Main/Main';
 
 // 용우's 컴포넌트
-import LoginYongwoo from './pages/yongwoo/Login/Login';
-import MainYongwoo from './pages/yongwoo/Main/Main';
+import LoginYongwoo from './pages/yongwoo/Login/LoginYongwoo';
+import MainYongwoo from './pages/yongwoo/Main/MainYongwoo';
 
 // 유진's 컴포넌트
 import LoginYoujin from './pages/youjin/Login/Login';
