+++ src/pages/yongwoo/Main/FeedsPictogram/FeedsPictogram.css
@@ -1,0 +1,36 @@
+.feedsPictogram {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-pack: justify;
+      -ms-flex-pack: justify;
+          justify-content: space-between;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+  width: 570px;
+  height: 40px;
+  margin: 0 auto;
+  padding-top: 10px;
+}
+
+.feedsPictogram .feedsPictogramLeft {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-pack: justify;
+      -ms-flex-pack: justify;
+          justify-content: space-between;
+  width: 100px;
+}
+
+.feedsPictogram .feedsPictogramLeft img {
+  width: 24px;
+  cursor: pointer;
+}
+
+.feedsPictogram .saveIcon {
+  width: 24px;
+  cursor: pointer;
+}
+/*# sourceMappingURL=FeedsPictogram.css.map */
