+++ src/pages/yongwoo/Main/FeedsPictogram/FeedsPictogram.scss
@@ -1,0 +1,24 @@
+@import '../../../../styles/variables.scss';
+
+.feedsPictogram {
+  @include flexBetween;
+  width: 570px;
+  height: 40px;
+  margin: 0 auto;
+  padding-top: 10px;
+
+  .feedsPictogramLeft {
+    display: flex;
+    justify-content: space-between;
+    width: 100px;
+
+    img {
+      width: 24px;
+      cursor: pointer;
+    }
+  }
+  .saveIcon {
+    width: 24px;
+    cursor: pointer;
+  }
+}
