--- src/pages/yongwoo/Main/Main.scss
@@ -1,324 +1,0 @@
-body {
-  padding-top: 40px;
-}
-$information-font-size: 11px;
-$information-font-color: #c7c7c7;
-$font-size: 14px;
-$border-gray: 1px solid rgb(219, 219, 219);
-
-@mixin flex-only-between {
-  display: flex;
-  justify-content: space-between;
-}
-
-@mixin flex-colums {
-  display: flex;
-  flex-direction: column;
-}
-@mixin flex-space-between {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.main {
-  display: flex;
-  margin-top: 300px;
-}
-
-.feedsContainer {
-  @include flex-only-between;
-  margin: 0 auto;
-  margin-right: 50px;
-
-  .feed {
-    @include flex-colums;
-    width: 614px;
-    margin-bottom: 30px;
-    border: 1px solid rgba(var(--b6a, 219, 219, 219), 1);
-
-    .photoWrap {
-      width: 612px;
-
-      img {
-        width: 100%;
-      }
-    }
-    .goodWrap {
-      display: flex;
-      align-items: center;
-      width: 570px;
-      margin: 0 auto;
-
-      p {
-        margin-left: 5px;
-        padding: 10px 0;
-        color: #262626;
-        font-size: $font-size;
-        font-weight: bold;
-      }
-
-      .goodImg {
-        width: 20px;
-        height: 20px;
-        border-radius: 10px;
-      }
-    }
-
-    .textWrap {
-      @include flex-colums;
-      width: 570px;
-      margin: 0 auto;
-      font-size: 14px;
-
-      .text {
-        color: #00376b;
-      }
-
-      p {
-        margin-top: 3px;
-        line-height: 20px;
-      }
-
-      .commentView {
-        color: #8e8e8e;
-      }
-
-      .time {
-        color: #8e8e8e;
-        font-size: 10px;
-      }
-
-      span {
-        font-size: $font-size;
-      }
-    }
-  }
-}
-
-.profileWrap {
-  @include flex-space-between;
-  width: 570px;
-  height: 60px;
-  margin: 0 auto;
-  padding: 20px 0;
-
-  .logoWrap {
-    display: flex;
-    align-items: center;
-
-    h2 {
-      margin-left: 5px;
-      font-size: $font-size;
-    }
-
-    i {
-      cursor: pointer;
-    }
-
-    img {
-      width: 32px;
-      height: 32px;
-      background-color: black;
-      border-radius: 16px;
-    }
-  }
-}
-
-.feedsPictogram {
-  @include flex-space-between;
-  width: 570px;
-  height: 40px;
-  margin: 0 auto;
-  padding-top: 10px;
-
-  .feedsPictogramLeft {
-    display: flex;
-    justify-content: space-between;
-    width: 100px;
-
-    img {
-      width: 24px;
-      cursor: pointer;
-    }
-  }
-  .saveIcon {
-    width: 24px;
-    cursor: pointer;
-  }
-}
-
-i.fas.fa-heart.redHeart {
-  display: inline-block;
-}
-i.fas.fa-heart.blackHeart {
-  display: inline-block;
-}
-
-.comment {
-  @include flex-space-between;
-
-  .commentNameBold {
-    font-weight: bold;
-  }
-
-  span.commentText {
-    font-weight: normal;
-  }
-
-  .heart {
-    width: 15px;
-    cursor: pointer;
-  }
-
-  .heartColor {
-    display: none;
-    width: 15px;
-    cursor: pointer;
-  }
-}
-
-.commentStart {
-  i {
-    font-size: 15px;
-
-    .fa-trash-alt {
-      display: none;
-      color: gray;
-      cursor: pointer;
-    }
-
-    .fa-heart {
-      display: none;
-      cursor: pointer;
-    }
-
-    .colorHeart {
-      display: none;
-      color: #f44336;
-      cursor: pointer;
-    }
-  }
-}
-
-.recommendContainer {
-  .mainRight {
-    width: 293px;
-    height: 460px;
-
-    .myInfoWrap {
-      @include flex-space-between;
-
-      .myInfoPhotoWrap {
-        @include flex-only-between;
-
-        img {
-          width: 56px;
-          height: 56px;
-          border-radius: 28px;
-          background-color: black;
-        }
-
-        .myInfo {
-          @include flex-colums;
-          justify-content: center;
-          margin-left: 18px;
-
-          h2 {
-            font-size: $font-size;
-          }
-
-          p {
-            margin-top: 2px;
-            color: #8e8e8e;
-            font-size: $font-size;
-          }
-        }
-      }
-
-      .infoBlue {
-        color: #0095f6;
-        font-size: 12px;
-        font-weight: bold;
-      }
-    }
-
-    .recommendTitle {
-      @include flex-only-between;
-      margin-top: 30px;
-
-      .grayColor {
-        color: #8e8e8e;
-        font-weight: bold;
-        font-size: $font-size;
-      }
-
-      .blackColor {
-        color: #262626;
-        font-size: 12px;
-        font-weight: 600;
-      }
-    }
-  }
-
-  .recommendWrap {
-    @include flex-space-between;
-    margin-top: 16px;
-
-    .recommend {
-      display: flex;
-
-      .recommendPhoto {
-        width: 32px;
-        height: 32px;
-        border: $border-gray;
-        border-radius: 16px;
-      }
-    }
-
-    .recommendText {
-      margin-left: 13px;
-
-      h4 {
-        color: #262626;
-        font-size: $font-size;
-      }
-
-      p {
-        margin-top: 3px;
-        color: #8e8e8e;
-        font-size: 12px;
-      }
-    }
-    .follow {
-      color: #0095f6;
-      font-size: 12px;
-      font-weight: bold;
-    }
-  }
-}
-
-.information {
-  .informatioinList {
-    display: flex;
-    align-items: center;
-    justify-content: start;
-    flex-wrap: wrap;
-    margin-top: 37px;
-    color: $information-font-color;
-    font-size: $information-font-size;
-
-    li {
-      padding-right: 8px;
-      line-height: 17px;
-      list-style: none;
-    }
-  }
-
-  p {
-    margin-top: 25px;
-    color: $information-font-color;
-    font-size: $information-font-size;
-    font-weight: 400;
-  }
-}
