+++ src/pages/yongwoo/Main/MainYongwoo.scss
@@ -1,0 +1,12 @@
+@import '../../../styles/variables.scss';
+
+body {
+  padding-top: 40px;
+}
+
+.mainYongwoo {
+  .main {
+    display: flex;
+    margin-top: 300px;
+  }
+}
