+++ src/pages/yongwoo/Main/ProfileWrap/ProfileWrap.css
@@ -1,0 +1,41 @@
+.profileWrap {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-pack: justify;
+      -ms-flex-pack: justify;
+          justify-content: space-between;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+  width: 570px;
+  height: 60px;
+  margin: 0 auto;
+  padding: 20px 0;
+}
+
+.profileWrap .logoWrap {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+}
+
+.profileWrap .logoWrap h2 {
+  margin-left: 5px;
+  font-size: 14px;
+}
+
+.profileWrap .logoWrap i {
+  cursor: pointer;
+}
+
+.profileWrap .logoWrap img {
+  width: 32px;
+  height: 32px;
+  background-color: black;
+  border-radius: 16px;
+}
+/*# sourceMappingURL=ProfileWrap.css.map */
