+++ src/pages/yongwoo/Main/RecommendContainer/RecommendContainer.scss
@@ -1,0 +1,125 @@
+@import '../../../../styles/variables.scss';
+
+.recommendContainer {
+  .mainRight {
+    width: 293px;
+    height: 460px;
+
+    .myInfoWrap {
+      @include flexBetween;
+
+      .myInfoPhotoWrap {
+        display: flex;
+        justify-content: space-between;
+
+        img {
+          width: 56px;
+          height: 56px;
+          border-radius: 28px;
+          background-color: black;
+        }
+
+        .myInfo {
+          display: flex;
+          flex-direction: column;
+          justify-content: center;
+          margin-left: 18px;
+
+          h2 {
+            font-size: 14px;
+          }
+
+          p {
+            margin-top: 2px;
+            color: #8e8e8e;
+            font-size: 14px;
+          }
+        }
+      }
+
+      .infoBlue {
+        color: #0095f6;
+        font-size: 12px;
+        font-weight: bold;
+      }
+    }
+
+    .recommendTitle {
+      display: flex;
+      flex-direction: column;
+      margin-top: 30px;
+
+      .grayColor {
+        color: #8e8e8e;
+        font-weight: bold;
+        font-size: 14px;
+      }
+
+      .blackColor {
+        color: #262626;
+        font-size: 12px;
+        font-weight: 600;
+      }
+    }
+  }
+
+  .recommendWrap {
+    @include flexBetween;
+    margin-top: 16px;
+
+    .recommend {
+      display: flex;
+
+      .recommendPhoto {
+        width: 32px;
+        height: 32px;
+        border: 1px solid #dbdbdb;
+        border-radius: 16px;
+      }
+    }
+
+    .recommendText {
+      margin-left: 13px;
+
+      h4 {
+        color: #262626;
+        font-size: 14px;
+      }
+
+      p {
+        margin-top: 3px;
+        color: #8e8e8e;
+        font-size: 12px;
+      }
+    }
+    .follow {
+      color: #0095f6;
+      font-size: 12px;
+      font-weight: bold;
+    }
+  }
+  .information {
+    .informatioinList {
+      display: flex;
+      align-items: center;
+      justify-content: start;
+      flex-wrap: wrap;
+      margin-top: 37px;
+      color: #c7c7c7;
+      font-size: 11px;
+
+      li {
+        padding-right: 8px;
+        line-height: 17px;
+        list-style: none;
+      }
+    }
+
+    p {
+      margin-top: 25px;
+      color: #c7c7c7;
+      font-size: 11px;
+      font-weight: 400;
+    }
+  }
+}
