+++ src/pages/yongwoo/Main/TextWrap/TextWrap.scss
@@ -1,0 +1,85 @@
+@import '../../../../styles/variables.scss';
+
+.textWrap {
+  display: flex;
+  flex-direction: column;
+  width: 570px;
+  margin: 0 auto;
+  font-size: 14px;
+
+  .text {
+    color: #00376b;
+  }
+
+  p {
+    margin-top: 3px;
+    line-height: 20px;
+  }
+
+  .comment {
+    @include flexBetween;
+
+    .commentNameBold {
+      font-weight: bold;
+
+      .commentText {
+        font-weight: normal;
+      }
+    }
+
+    .heart {
+      width: 15px;
+      cursor: pointer;
+    }
+
+    .heartColor {
+      display: none;
+      width: 15px;
+      cursor: pointer;
+    }
+  }
+
+  .commentStart {
+    i {
+      font-size: 15px;
+
+      .fa-trash-alt {
+        display: none;
+        color: gray;
+        cursor: pointer;
+      }
+
+      .fa-heart {
+        display: none;
+        cursor: pointer;
+      }
+
+      .colorHeart {
+        display: none;
+        color: #f44336;
+        cursor: pointer;
+      }
+    }
+  }
+
+  .commentView {
+    color: #8e8e8e;
+  }
+
+  .time {
+    color: #8e8e8e;
+    font-size: 10px;
+  }
+
+  span {
+    font-size: 14px;
+  }
+
+  i.fas.fa-heart.redHeart {
+    display: inline-block;
+  }
+
+  i.fas.fa-heart.blackHeart {
+    display: inline-block;
+  }
+}
