+++ src/styles/variables.css
@@ -1,0 +1,1 @@
+/* No CSS *//*# sourceMappingURL=variables.css.map */
